Settings Menu
Well, the settings menu exists, sure, but uh, nope.
Known bugs so far:
When you change the resolution from default, enter a level, and then get back to the main menu through any means, the resolution resets to default
When the fullscreen box is unchecked (making the game no longer fullscreen), the game is closed, and then the game is reopened, the fullscreen box is still checked (easy fix)
There's literally no way to tell if the volume is actually working other than the fact that it should, due to the fact that there is no sound in this game (yet).